SOPs/pdf-process/requirements.txt
python-dotenv
requests
pymupdf
google-genai
pillow